requirements.txt
streamlit
psycopg2-binary
bcrypt
opencv-python
face-recognition
numpy
pillow
